Remove unnecessary async, fix type definitions
