+++ gmqtt/__init__.py
@@ -1,17 +1,19 @@
 import datetime
 
-from .client import Client, Message, Subscription
+from .client import Client, Message
+from .subscription import Subscription
 from .mqtt import constants
 from .mqtt.protocol import BaseMQTTProtocol
 from .mqtt.handler import MQTTConnectError
 
 __author__ = "Mikhail Turchunovich"
-__email__ = 'mitu@gurtam.com'
+__email__ = 'dyez@gurtam.team'
 __copyright__ = ("Copyright 2013-%d, Gurtam; " % datetime.datetime.now().year,)
 
 __credits__ = [
     "Mikhail Turchunovich",
-    "Elena Shylko"
+    "Elena Shylko",
+    "Dmitry Yezepchik"
 ]
 __version__ = "0.7.0"
