+++ gmqtt/client.py
@@ -3,13 +3,13 @@ import json
 import logging
 import uuid
 from functools import partial
-from typing import Sequence, Union
 
 from .mqtt.connection import MQTTConnection
 from .mqtt.constants import UNLIMITED_RECONNECTS, MQTTv50
 from .mqtt.handler import MqttPackageHandler
 from .mqtt.utils import ConnectionState
 from .storage import PersistentStorage
+from .subscription import SubscriptionsHandlerMixin
 
 
 class Message:
@@ -38,126 +38,6 @@ class Message:
 
         if self.payload_size > 268435455:
             raise ValueError("Payload too large.")
-
-
-class Subscription:
-    def __init__(
-        self,
-        topic,
-        qos=0,
-        no_local=False,
-        retain_as_published=False,
-        retain_handling_options=0,
-        subscription_identifier=None,
-    ):
-        self.topic = topic
-        self.qos = qos
-        self.no_local = no_local
-        self.retain_as_published = retain_as_published
-        self.retain_handling_options = retain_handling_options
-
-        self.mid = None
-        self.acknowledged = False
-
-        # this property can be used only in MQTT5.0
-        self.subscription_identifier = subscription_identifier
-
-
-class SubscriptionsHandlerMixin:
-    def __init__(self):
-        self.subscriptions = []
-        # defined by a main class
-        self._connection = None
-
-    def update_subscriptions_with_subscription_or_topic(
-        self,
-        subscription_or_topic,
-        qos,
-        no_local,
-        retain_as_published,
-        retain_handling_options,
-        kwargs,
-    ):
-
-        sentinel = object()
-        subscription_identifier = kwargs.get("subscription_identifier", sentinel)
-
-        if isinstance(subscription_or_topic, Subscription):
-
-            if subscription_identifier is not sentinel:
-                subscription_or_topic.subscription_identifier = subscription_identifier
-
-            subscriptions = [subscription_or_topic]
-        elif isinstance(subscription_or_topic, (tuple, list)):
-
-            if subscription_identifier is not sentinel:
-                for sub in subscription_or_topic:
-                    sub.subscription_identifier = subscription_identifier
-
-            subscriptions = subscription_or_topic
-        elif isinstance(subscription_or_topic, str):
-
-            if subscription_identifier is sentinel:
-                subscription_identifier = None
-
-            subscriptions = [
-                Subscription(
-                    subscription_or_topic,
-                    qos=qos,
-                    no_local=no_local,
-                    retain_as_published=retain_as_published,
-                    retain_handling_options=retain_handling_options,
-                    subscription_identifier=subscription_identifier,
-                )
-            ]
-        else:
-            raise ValueError(
-                "Bad subscription: must be string or Subscription or list of Subscriptions"
-            )
-        self.subscriptions.extend(subscriptions)
-        return subscriptions
-
-    def _remove_subscriptions(self, topic: Union[str, Sequence[str]]):
-        if isinstance(topic, str):
-            self.subscriptions = [s for s in self.subscriptions if s.topic != topic]
-        else:
-            self.subscriptions = [s for s in self.subscriptions if s.topic not in topic]
-
-    def subscribe(
-        self,
-        subscription_or_topic: Union[str, Subscription, Sequence[Subscription]],
-        qos=0,
-        no_local=False,
-        retain_as_published=False,
-        retain_handling_options=0,
-        **kwargs
-    ):
-
-        # Warn: if you will pass a few subscriptions objects, and each will be have different
-        # subscription identifier - the only first will be used as identifier
-        # if only you will not pass the identifier in kwargs
-
-        subscriptions = self.update_subscriptions_with_subscription_or_topic(
-            subscription_or_topic,
-            qos,
-            no_local,
-            retain_as_published,
-            retain_handling_options,
-            kwargs,
-        )
-        return self._connection.subscribe(subscriptions, **kwargs)
-
-    def resubscribe(self, subscription: Subscription, **kwargs):
-        # send subscribe packet for subscription,that's already in client's subscription list
-        if "subscription_identifier" in kwargs:
-            subscription.subscription_identifier = kwargs["subscription_identifier"]
-        elif subscription.subscription_identifier is not None:
-            kwargs["subscription_identifier"] = subscription.subscription_identifier
-        return self._connection.subscribe([subscription], **kwargs)
-
-    def unsubscribe(self, topic: Union[str, Sequence[str]], **kwargs):
-        self._remove_subscriptions(topic)
-        return self._connection.unsubscribe(topic, **kwargs)
 
 
 class Client(SubscriptionsHandlerMixin):
@@ -371,8 +251,7 @@ class Client(SubscriptionsHandlerMixin):
 
     def _remove_message_from_queue(self, mid):
         self._logger.debug("[Client] remove message. mid: %s", mid)
-        future = asyncio.ensure_future(self._persistent_storage.remove_message_by_mid(mid))
-        future.add_done_callback(partial(self._handle_exception_in_future, msg="remove message"))
+        self._persistent_storage.remove_message_by_mid(mid)
 
     @property
     def is_connected(self):
@@ -382,23 +261,23 @@ class Client(SubscriptionsHandlerMixin):
     async def _resend_qos_messages(self):
         await self._connack_received.wait()
 
-        if await self._persistent_storage.is_empty:
+        if self._persistent_storage.is_empty:
             self._logger.debug("[Client] QoS queue is empty, nothing to replay")
             return
         elif self._connection.is_closing():
             self._logger.warning(
                 "[Client] transport already closing, skipping replay of %s message(s) — "
                 "next reconnect will retry",
-                len(await self._persistent_storage.get_all()),
+                len(self._persistent_storage.get_all()),
             )
             return
         else:
-            msgs = await self._persistent_storage.get_all()
+            msgs = self._persistent_storage.get_all()
             self._logger.debug(
                 "[Client] replaying %s inflight message(s)", len(msgs)
             )
 
-            await self._persistent_storage.clear()
+            self._persistent_storage.clear()
 
             for mid, package in msgs:
                 try:
@@ -411,8 +290,8 @@ class Client(SubscriptionsHandlerMixin):
 
                 self._persistent_storage.push_message(mid, package)
 
-    async def _clear_resend_qos_queue(self):
-        await self._persistent_storage.clear()
+    def _clear_resend_qos_queue(self):
+        self._persistent_storage.clear()
 
     def set_auth_credentials(self, username, password=None):
         self._username = username.encode()
